try to build package after build

+++ buildOrthancOSX.sh
@@ -2,50 +2,58 @@
 
 set -ex
 
-# CHANGE_VERSION_OSX
-VERSION=22.3.0
+is_tag=${1:-false}
+branch_tag_name=${2:-22.3.0}       # CHANGE_VERSION_OSX
+download_from_buildbot=${3:-true}  # if false, we consider that binaries are already in /tmp/Orthanc-OSX-$branch_tag_name
+
 
 URL='https://alain:koo4oCah@buildbot.orthanc-server.com/artifacts/Binaries/'
 TARGET='/tmp/'
-FOLDER=Orthanc-OSX-$VERSION
+FOLDER=Orthanc-OSX-$branch_tag_name
 
 # https://stackoverflow.com/a/4774063/881731
 SCRIPTPATH="$( cd "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
 
-rm -rf ${TARGET}/${FOLDER} 
-mkdir ${TARGET}/${FOLDER}
+if [[ $download_from_buildbot == "true" ]]; then
+    rm -rf ${TARGET}/${FOLDER} 
+    mkdir ${TARGET}/${FOLDER}
+fi
+
 cd ${TARGET}/${FOLDER}
 
-# Resources
-cp ${SCRIPTPATH}/orthancBuildResources/configOSX.json .
-cp ${SCRIPTPATH}/orthancBuildResources/readmeOSX.txt readme.txt
-cp ${SCRIPTPATH}/orthancBuildResources/startOrthanc.command .
-cp ${SCRIPTPATH}/WindowsInstaller/Resources/ca-certificates.crt .
+if [[ $download_from_buildbot == "true" ]]; then
 
-# Orthanc core
-wget "${URL}/Orthanc 1.10.1 - OS X Release/Orthanc"
-wget "${URL}/Orthanc 1.10.1 - OS X Release/libModalityWorklists.dylib"
-wget "${URL}/Orthanc 1.10.1 - OS X Release/libServeFolders.dylib"
-wget "${URL}/Orthanc 1.10.1 - OS X Release/libConnectivityChecks.dylib"
+    # Resources
+    cp ${SCRIPTPATH}/orthancBuildResources/configOSX.json .
+    cp ${SCRIPTPATH}/orthancBuildResources/readmeOSX.txt readme.txt
+    cp ${SCRIPTPATH}/orthancBuildResources/startOrthanc.command .
+    cp ${SCRIPTPATH}/WindowsInstaller/Resources/ca-certificates.crt .
 
-chmod +x ./Orthanc
+    # Orthanc core
+    wget "${URL}/Orthanc 1.10.1 - OS X Release/Orthanc"
+    wget "${URL}/Orthanc 1.10.1 - OS X Release/libModalityWorklists.dylib"
+    wget "${URL}/Orthanc 1.10.1 - OS X Release/libServeFolders.dylib"
+    wget "${URL}/Orthanc 1.10.1 - OS X Release/libConnectivityChecks.dylib"
 
-# Plugins
-wget "${URL}/DicomWeb 1.7 - OS X Release/libOrthancDicomWeb.dylib"
-wget "${URL}/MySQL 4.3 - OS X Release/libOrthancMySQLIndex.dylib"
-wget "${URL}/MySQL 4.3 - OS X Release/libOrthancMySQLStorage.dylib"
-wget "${URL}/OsimisWebViewer 1.4.2 - OS X Release/libOsimisWebViewer.dylib"
-wget "${URL}/PostgreSQL 4.0 - OS X Release/libOrthancPostgreSQLIndex.dylib"
-wget "${URL}/PostgreSQL 4.0 - OS X Release/libOrthancPostgreSQLStorage.dylib"
-wget "${URL}/Transfers 1.0 - OS X Release/libOrthancTransfers.dylib"
-wget "${URL}/WebViewer 2.8 - OS X Release/libOrthancWebViewer.dylib"
-wget "${URL}/Stone 2.3 - OS X Release/libStoneWebViewer.dylib"
-wget "${URL}/Gdcm 1.5 - OS X Release/libOrthancGdcm.dylib"
-wget "${URL}/WSI 1.1 - OS X Release/libOrthancWSI.dylib"
-wget "${URL}/Odbc 1.1 - OS X Release/libOrthancOdbcIndex.dylib"
-wget "${URL}/Odbc 1.1 - OS X Release/libOrthancOdbcStorage.dylib"
-wget "${URL}/Tcia 1.1 - OS X Release/libOrthancTcia.dylib"
-wget "${URL}/FolderIndexer 1.0 - OS X Release/libOrthancIndexer.dylib"
+    chmod +x ./Orthanc
+
+    # Plugins
+    wget "${URL}/DicomWeb 1.7 - OS X Release/libOrthancDicomWeb.dylib"
+    wget "${URL}/MySQL 4.3 - OS X Release/libOrthancMySQLIndex.dylib"
+    wget "${URL}/MySQL 4.3 - OS X Release/libOrthancMySQLStorage.dylib"
+    wget "${URL}/OsimisWebViewer 1.4.2 - OS X Release/libOsimisWebViewer.dylib"
+    wget "${URL}/PostgreSQL 4.0 - OS X Release/libOrthancPostgreSQLIndex.dylib"
+    wget "${URL}/PostgreSQL 4.0 - OS X Release/libOrthancPostgreSQLStorage.dylib"
+    wget "${URL}/Transfers 1.0 - OS X Release/libOrthancTransfers.dylib"
+    wget "${URL}/WebViewer 2.8 - OS X Release/libOrthancWebViewer.dylib"
+    wget "${URL}/Stone 2.3 - OS X Release/libStoneWebViewer.dylib"
+    wget "${URL}/Gdcm 1.5 - OS X Release/libOrthancGdcm.dylib"
+    wget "${URL}/WSI 1.1 - OS X Release/libOrthancWSI.dylib"
+    wget "${URL}/Odbc 1.1 - OS X Release/libOrthancOdbcIndex.dylib"
+    wget "${URL}/Odbc 1.1 - OS X Release/libOrthancOdbcStorage.dylib"
+    wget "${URL}/Tcia 1.1 - OS X Release/libOrthancTcia.dylib"
+    wget "${URL}/FolderIndexer 1.0 - OS X Release/libOrthancIndexer.dylib"
+fi
 
 # Create the archive
 cd ${TARGET}
